Qnx: Refactor BlackBerryApplicationRunner class
This patch refactors the BlackBerryApplicationRunner class and
instantiates two log processes using slog2info and tail commands.
The latter is used to get error logs if the application fails to launch.

Slog2info display logs only when the app is correctly launched.
If the app failed to launch (failed to load shared libraries... etc)
the error logs are then written in /accounts/.../appData/logs/log.

Change-Id: Ie2804ba7a3f67e40ece1e6973cfbcb2ee11ae0b0
Reviewed-by: Tobias Nätterlund <tobias.naetterlund@kdab.com>
Reviewed-by: Nicolas Arnaud-Cormos <nicolas@kdab.com>