Attach to unstarted Application
This patch adds a dialog that can be useful in situations where you have
a script-based test tool, that starts executable and executes test on
it.

This dialog allows user to specify executable and to begin watching
process, where dialog simply waits for executable to be started. As soon
as executable is found, QtCreator will attach to it.

also, since script-based test can constantly start-stop executable. as
soon as it is detected that executable is down, another waiting process
starts where dialog waits for executable to come up again.

Change-Id: I3bd62f656d34f8dff5641f6747ec12d97cbae2a4
Reviewed-by: hjk <hjk121@nokiamail.com>