Autoindent: Don't reindent on every autotext insert.
Currently the line is completely reintented whenever an autoText
insert is being done (like inserting a matching paren). This is un-
necessary as well as actively harmful.

Unnecessary, because the electric character mechanism is already
supposed to trigger reindents where necessary.

Harmful, because the reindent method that is used does not take care
to preserve explicit user indentation. The electric character reindent
does that (by only changing indentation if the actual current indent
is the same as the previously expected indent).

Change-Id: I19db853b6c30792f867baad919bd4eb9b4c23361
Reviewed-by: Nikolai Kosjar <nikolai.kosjar@digia.com>