CppEditor: fix follow symbol under cursor in SIGNAL/SLOT macros
If the cursor was on the first character of a SIGNAL/SLOT name in a
QObject::connect call, the detection in findLinkAt failed.
Token::end() is the first position behind the token, so we must check
if positionInBlock is less than tk.end().
An extra branch has been added to handle the case where the cursor is
on the opening parenthesis after the identifier within a SIGNAL/SLOT
macro. This worked before, because of the "fuzziness" of the former
condition.

Task-number: QTCREATORBUG-10264

Change-Id: I2122bc2c13af1eb6c37dba9134d2b5ab931f80a9
Reviewed-by: Nikolai Kosjar <nikolai.kosjar@digia.com>