AndroidExtraLibraryListModel: Fix crash on removing libs
Steps to reproduce the crash, start with the list a b a c, select the
second a and the c. Click on remove. The removeEntries will find that it
can remove two consecutive entries in one beginRemoveRows/endRemoveRows,
but will wrongly remove them starting at the first a.

The fix is too simply order the modelindexes in descendeding order such
that removing entries does not modify the rows.

Change-Id: I4be349f4bab8137075da0d8dfcef24f10dc25f92
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@digia.com>