TextEditor: Fix displaying function proposals
In commit 3a64f8a the condition for displaying a replacement proposal
was strengthened by making it only depend on the current proposal
m_proposal. Since FuntionHintProposals were fragile, the automatic
proposal timer was restarted in notifyChange() and as consequence the
proposal was reset.

Fixed by declaring that FunctionHintProposal are not fragile.

Task-number: QTCREATORBUG-10283
Change-Id: I60828d733b781279be51228bd10d49f828aebe2b
Reviewed-by: Erik Verbruggen <erik.verbruggen@digia.com>