QmlDesigner: Support for enums in alias properties
Enums in combination with alias properties were not supported.
The QML Engine does not distinguish between bindings
and enums in this case.
We want to distinguish between bindings/expressions and literal
enums. This is not possible with the current code model, so we use
a list of hardcoded enum keys for Qt Quick.

There a several issues we want to properly fix in the future.

*  Make the model aware of enums. Currently enums are strings
   (the key) in the model. This has historical reasons. Instead
   we should use a custom type registered to QVariant that handles
   the scope, key and integer value.

*  The code model should be fully aware of enums even in the alias case.

Do not integrate this patch to 3.0 or master.

Task-number: QTCREATORBUG-10114
Change-Id: I780a201202f0949b1e2eb2f2525ddb031e1cd18b
Reviewed-by: Marco Bubke <marco.bubke@digia.com>