QmlDesigner.MetaInfo: fix handling of enum scopes
The codel model does not provide enum scopes at the moment.
As a fallback we use class names, but this fails if we use
the original cpp class name from the <cpp> package.
For this reason we always try to find an export
with a package different from <cpp>.
As a final solution the code model should be able to
provide the enum scopes.

Task-number: QTCREATORBUG-10114
Change-Id: I15396c83590426ab0b9b55c8646a89c8d5712683
Reviewed-by: Fawzi Mohamed <fawzi.mohamed@digia.com>