Android: Remove bestMatch calls in fallback code in buildTargetSDK
besdtMatch finds a ndk version that is at most the targetApi passed into it.
Since we are passing in the fallback and lowest possible api level that qt
supports, the function would return at most the lowest supported version.
That's not very useful and cannot have worked, so simpy return the fallback
version directly.

Change-Id: I5e4757639099af4d207f928b0335d28efd75a02b
Reviewed-by: BogDan Vatra <bogdan@kde.org>