Increase preference dialog height by 10% on Mac
Widgets and layouts take more space on Mac than on the other platforms,
and that results in developers designing pages that are too heigh for
Mac even though they fit on other platforms. Since redesigning
preference pages so they also fit on Mac all the time is not ideal,
simply account for the larger needed space.

Change-Id: Ib8483713f562436bc9a86063f345ae4e44473608
Reviewed-by: Erik Verbruggen <erik.verbruggen@digia.com>