Don't make the output pane get taller when switching tabs.
A code review change broke the logic. Revise to avoid this bug.
Load the saved height once at startup and save it at app close.
Track changes to the height as they happen.
This ensures the output pane does not get taller if you shrink
it and then switch tabs.

Task-number: QTCREATORBUG-8877
Change-Id: If0d0d0d45f2944477003eb407e64567441bad414
Reviewed-by: Eike Ziller <eike.ziller@digia.com>