DebuggerItem: Make sure DebuggerItems have an Id
The default constructor now sets an id on debuggerItem. fromMap was
removed and a new constructor (taking a const QVariantMap &) was
added in place of that.

There are no more friends on the DebuggerItem class since those were
only necessary to make sure the item has an Id when it gets added.

Change-Id: Ia1a6c9ffea67a8e0a1e5685ef93f67df8686d4c9
Reviewed-by: hjk <hjk121@nokiamail.com>
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>