QbsProjectManager: Fix display of command line in step widgets.
The command line equivalent of the respective options often gets too
long for the summary, which means that the end of the string is cut off.
Not only does that look ugly, it also hides information that is not
trivially available otherwise, such as the full path to the install root.
This patch therefore adds a text edit field that is guaranteed to be
able to display the complete command line. This is also consistent
with what the qmake step does.

Change-Id: Ic8d20e484e9cfe4980ea9eff05843fef1012f067
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>