Qnx: Cascades project template is obsolete
Qt Creator QTCREATORBUG-10375

The old Cascades template application was more sample than template
and it was simplified.

The application UI object name doesn't follow the project name anymore.
This change was introduced because C++ class name rules are more
restrictive than the project ones. More than that, it is harder
to start with some test project and then refractor it to some real one.
There is no gain to have UI object class name same as the enclosing
project one.

The application icon added to the bar-descriptor.xml

Change-Id: I198f61a631fa2c758f3e8ac377c06ce7560cc911
Reviewed-by: Nicolas Arnaud-Cormos <nicolas@kdab.com>