SDKtool: Add support for debugger.xml
Add support to add/remove debugger entries to debuggers.xml

This is in addition to the existing mode of writing the debugger
path/debugger engine into the kit. This works reasonably well, but
the new mode offers better control as how the debugger is called, etc.

addKit also got a --debuggerid switch to set the debugger id instead
of using the old configuration format.

Change-Id: Id4be0366d2c126252419cf1ba87cd0e117204e54
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>