GccToolChain: Improve detection of 64bit/32bit
No longer use the undocumented __X86_64 macro to decide on the
word width of a tool chain target. Use __SIZEOF_SIZE_T__ instead.
That is documented to be set and should also be applicable outside
of the intel x86 world.

Some versions of mingw do not define __X86_64 on 64bit intel
machines and thus are not properly detected.

Update unit tests that try to guess the ABI to use the new macro
instead and fix the fallout.

Side effect: Some abis are detected in a different sequence now.
That should not change the tool chains though as the ABI that is actually
used is stored extra and does not referrence that list.

Task-number: QTCREATORBUG-10584
Change-Id: Ia9e5909cdff3b057ef74b24c0f2d654a14253c21
Reviewed-by: Daniel Teske <daniel.teske@digia.com>