Qnx: executing instead of parsing bbndk-env script file
QnxUtils::qnxEnvironmentFromNdkFile() function was parsing the content
of the bbndk-env.sh/bat script file to resolve environment for BB10 tools.

The content of the script file is not documented and is a subject
of heavy and frequent changes.

Now QnxUtils::qnxEnvironmentFromNdkFile() function executes the script file
and dumps the important env. variables that are parsed standard way.

Change-Id: Ie7636054d62ec62b32cd3ab3d93a59b7d6c0882a
Reviewed-by: Tobias Nätterlund <tobias.naetterlund@kdab.com>
Reviewed-by: Nicolas Arnaud-Cormos <nicolas@kdab.com>