Android: getAbis prevent endless loop
We need to check all ro.product.cpu.abi# properties,
but there's a limited amount of those. So only check up to 5 to ensure
that the loop always terminates. On irc a user reported that on his
device adb shell getprop returns  "/bin/sh: getprop: not found", so
this loop was never terminated.

Change-Id: I819b28cdbedb0bf557ba58e8495a8da5eeb8dc8d
Reviewed-by: Mitch Curtis <mitch.curtis@digia.com>