Debugger: Disable theadname extraction (again)
This only affects MinGW where the feature triggered QTCREATORBUG-5200 in
the past. Right now the performance impact of c9c0f80 (fixing
QTCREATORBUG-10826) is unclear and no time to investigate. Trading the
feature for robustness seems in order. Power-users can remove the return
line in the script to get it back.

Change-Id: If5a30808b72969493e5f40dd0fcbafdcccec343e
Reviewed-by: Eike Ziller <eike.ziller@digia.com>
Reviewed-by: Orgad Shaneh <orgads@gmail.com>