Android: Show output of android create avd
The output is shown in the general message pane. Which is rather
suboptimal since both ways to create avds are via a modal dialog, thus
the output is not even accessible  easily. So far no one has reported
issues with failed avd creation. In case someone does, we now have the
output of the command.

Task-number: QTCREATORBUG-10601
Change-Id: I41c68d48fd317cd0e892a5ab3f5e2a5fa4bcbae9
Reviewed-by: BogDan Vatra <bogdan@kde.org>