Doc: use cross-linking to Qt module documentation
Replace links to "\l{http://qt-project.org/doc/"
with links to topic titles. Remove links where QDoc
creates them automatically based on the .index files.
Add the .index files for the modules that are discussed
in the Qt Creator Manual.

Note: the links will only be generated if the Qt used to
build the docs contains the linked modules and documentation
(.index file) has been generated for them.

Change-Id: Ibe624bf3773e7c854c03ebba4db406be0b4a7b90
Reviewed-by: Jerome Pasion <jerome.pasion@digia.com>
Reviewed-by: Eike Ziller <eike.ziller@digia.com>