Editor: Fix crash in the generic highighlighter.
Remove the TextBlockUserData base class BlockData and store the data of
this class inside a CodeFormatter base class. This class can be added to the
TextBlockUserData via setCodeFormatterData. Now we don't have to assume
that the user data is a specific base class of TextBlockUserData.

This removes a crash where the TextBlockUserData was created before the
highlighter could create and set his base class.

Task-number: QTCREATORBUG-10871

Change-Id: I167bdb68b9b1fecc64e4906bdad60bfbecb3bf47
Reviewed-by: Eike Ziller <eike.ziller@digia.com>