WelcomeScreen: Avoid loading of invalid pictures
Apparently the QML engine also loads pictures for invisible Image
objects. The pictures icons for examples and tutorials on the welcome
screen thus always try to load an extra video icon, which always fails.
By setting the URL to an empty string if the example or tutorial being
loaded is no video we can avoid that.

Change-Id: Ibafc11ed233f386bbbf1e7a4830fcb34bc1cd55d
Reviewed-by: Eike Ziller <eike.ziller@digia.com>
Reviewed-by: Thomas Hartmann <Thomas.Hartmann@digia.com>