Utils: Fix getting a case sensitive path on windows.
Until now we used GetShortPathName which relies on having 8.3 paths.
This short path names are converted into long names which also corrects
the case. The 8.3 paths can be manually deactivated via the fsutil. When
this is done GetShortPathName does not return a convertible path string.

If the case isn't corrected on windows we are opening multiple editors for
the same file. For example, when reaching a breakpoint when using cdb or
opening an entry from the issue panel generated by the msvc compile.

Task-number: QTCREATORBUG-11199
Change-Id: Ic04055685e0e9bbe0d45cde20563929b7370695d
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Eike Ziller <eike.ziller@digia.com>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>