Update cache size when pixmap size becomes known
The QML engine can reserve space in the pixmap cache (a QHash of
pointers) significantly before the pixmap is actually loaded. We can
only be sure about the actual size of the loaded image after it has been
loaded, though. This means we have to show pixmap cache updates in two
stages. One for "space reserved" and for "pixmap actually cached"

Change-Id: Ic7a7cd522e74d8b9b42dfbd6a9a0c4e9ffd6a401
Reviewed-by: Kai Koehne <kai.koehne@digia.com>