ClangCodeModel: Add qbs project file.
Not all logic present in the qmake project file for determining paths and
library names is replicated here: At the moment, all paths are derived from
the LLVM_INSTALL_DIR environment variable, and the library name is assumed
to be always "clang". Consolidation can be done when stabilizing the
logic.

Change-Id: Ib1ac8cf2465262e90325addb46f821549048a172
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>