QmlProfiler: Hide timline and show loading indicators when profiling
This is a temporary fix to work around the fact that the profiler's
loading indicator won't show on top of the timeline view. It's not
particularly beautiful but at least the UI doesn't just "hang" anymore.
A final solution has to be provided in qtdeclarative by allowing widgets
to be placed on top of QML scenes. In principle it's also possible to
change the loading indicators to be actual windows and by doing that have
them stay on top, but the result would be very messy.

Change-Id: Ic774b0792786d7d278da5ba0beda4531f0dba77d
Reviewed-by: Kai Koehne <kai.koehne@digia.com>