Mac: Prevent switching to discrete GPU on startup
By default, OpenGL-based apps will switch multi-GPU
Macs over to discrete graphics hardware.

It is possible to prevent this, and it makes sense
to do so for Qt Creator since its users will typically
be more concerned with power usage than raw OpenGL
performance.

Commit 8166a34f makes the necessary change to QtBase
and this commit enables the non-switching behavior
by adding the NSSupportsAutomaticGraphicsSwitching
key to the Info.plist

Change-Id: Iea6ef4b4922645f792cfb9b8b0b8860f5b2c0674
Reviewed-by: Eike Ziller <eike.ziller@digia.com>
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@digia.com>