QmlDesigner.PropertyEditor: Crash fix
There is no guarantee that the selection is still valid, when
changeExpression() is called. The signal might be triggered by an
EditingFinished() signal that was emitted after the selection was changed.

Just ignoring the change is not the ideal solution, but there is not much
more we can do.

Task-number: QTCREATORBUG-11249
Change-Id: I770dade680b0b8a052efd910b636829d1037e3ad
Reviewed-by: Robert Loehning <robert.loehning@digia.com>