QmlDesigner.PropertyEditor: fixing layout issues
There were serious layout issues in the property editor.
While not 100% reliable reproducible they were quite annoying (TASK).

The reason for the issues were the existence of several QtQuickViews and
changing visibility of items.
The pane for Item and the empty pane for no selection use different
QtQuickViews in a StackWidget. Somehow visibility of items is not always
updated properly when changing the QtQuickView in the stack.

Solution: Do not use visibility but stack order instead.
For the layout margins/targets all controls are now always visible, but
unused margins/targets get disabled.

Task-number: QTBUG-36141
Change-Id: Iec62b81796301aaf483606d9aca5a4ad3572e98b
Reviewed-by: Jan Arve Sæther <jan-arve.saether@digia.com>