SSH: Act on failing SFTP server.
So far we ignored crashes and unexpected exits of the remote SFTP
service under the assumption that the SSH server would catch these
itself and act accordingly. This is not the case, however: OpenSSH,
for instance, does not even realize if its sftp-server binary is
not present at all. As a result, Qt Creator waits indefinitely for
an SFTP operation to finish. Now we emit an error and close the
offending channel.

Task-number: QTCREATORBUG-10339
Change-Id: I132ed4a0098434a4cfce6056b964bd6363951fd7
Reviewed-by: Volker Vogelhuber <wiendl@web.de>
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>