Improve key handling for welcome screen
This makes the Projects/Examples/Tutorials buttons work
by keys in addition to mouse.

Instead of only respecting the mouse for button presses
use an ExclusiveGroup so that pressing space also changes
the mode (before pressing space on the Examples button
would change to example mode but not highlight the button).

Change-Id: Ib4704fc1bb09767b6b69e2a6364ceca4525a0676
Reviewed-by: Eike Ziller <eike.ziller@digia.com>