BlackBerry: Improve handling of Qt libraries when creating package
This lets the user choose between bundling the Qt libraries in the
bar package, use the libraries shipped with the device, or use
libraries deployed to the device by the user.

From the user's choice, the correct assets will be added to the bar
package, and the correct environment will be set up, so those does
not have to be pre-filled in the bar-descriptor.xml template any longer.

Task-number: QTCREATORBUG-11376
Change-Id: If44a9f9868d068f2ccbbeca3e25447496a336648
Reviewed-by: Nicolas Arnaud-Cormos <nicolas@kdab.com>