QNX: Add support for attaching to a running application
This adds a new menu item to the Debug -> Start Debugging menu,
which shows a processes dialog for the user, listing the processes
on the selected device. Clicking Attach to Application launches
pdebug on the device and then starts the gdb engine to attach
to the selected process.

This adds a DeviceTypeKitChooser class to ProjectExplorer, to
only have the KitChooser show kits configured for a specific
device type.

Change-Id: I1572dc76a6bdc4774dbf2e8cc96ca6bcf0f79a0a
Reviewed-by: hjk <hjk121@nokiamail.com>