Android Settings: Download buttons for Android Toolkit parts
Show little toolButtons next to the path input fields for SDK,
NDK, Ant and OpenJDK. The latter two only for Windows. These
buttons open the system Web browser with the according download
page.

We hope this saves the user a bit of time understanding what to
download where when setting up an Android development environ-
ment.

The download icon was drawn by me.

Change-Id: I0079059deca8d0f07d53b914900cb34b8f0489e0
Reviewed-by: Daniel Teske <daniel.teske@digia.com>