ProjectNodes: Do not try to sort directories not yet mapped
Do not try to sort directories that were not yet mapped into the
FlatModel. This avoids an assert where QModelIndex is out of range
(-1) because the data for that directory is not yet available in
the model.

Do not bother to map the data, just forget the request to sort in
this case as the data will be sorted when it gets mapped anyway.

Task-number: QBS-521
Change-Id: I19de7513405e6a0b84c988a2a86f763e3135811d
Reviewed-by: Daniel Teske <daniel.teske@digia.com>