Fix escape handling in find tool bar.
It was explicitly handling the escape key for "return to current
document find", which is wrong since the shortcut for "return to editor"
(which it replaces) can be changed.
The correct thing is to register an own action for "return to editor"
for the find tool bar, overriding the default behavior.

Change-Id: I77c690fa5921ce4022b8d2b38383668efd717875
Reviewed-by: hjk <hjk121@nokiamail.com>
Reviewed-by: Eike Ziller <eike.ziller@digia.com>